enhancement: abstract changeImportSpecifier as a utility method
